# server/config/swagger/kibopush.neweditor.yaml
...
swagger: '2.0'
...
info:
  version: v1
  title: KiboPush API
  description: |
    Welcome to the KiboPush API documentation! Our API provides flexible tools that can be used in the creation of new, outstanding projects.
    Please note that this documentation refers to the API 1.0 version. API 2.0 version is currently under construction and will be released soon.

    Build something great!

    ## Basic Usage
    All KiboPush API requests start with `https://app.kibopush.com/api/`

    The next segment of the URI path depends on the type of request. For example:

    `https://app.kibopush.com/api/users/`

    to get information of logged in user.

    ## Authentication
    In order for client application to call our API, he/she would have to attach the following information to the http headers:

      * app_id (the API key)
      * app_secret (the API key secret)


    ## Error Handling
    Errors are returned using standard HTTP error code syntax. In general, codes in the 2xx range indicate success, codes in the 4xx range indicate an error (bad or missing parameters, not authenticated etc.), and codes in the 5xx range indicate an error with KiboPush’s servers. Any additional info is included in the body of the return call, JSON-formatted.

  termsOfService: http://www.kibopush.com/privacy-policy

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
...
host: app.kibopush.com
basePath: /api
schemes:
...
produces:
...
consumes:
...
securityDefinitions:
  apiKey:
    type: apiKey
    in: header
    name: app_id
  apiSecret:
    type: apiKey
    in: header
    name: app_secret
security:
  - apiKey: []
    apiSecret: []

################################################################################
#                                           Paths                              #
################################################################################
paths:
...
  /templates/allPolls:
    get:
      description: Gives all the template polls
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: array
                items:
                  $ref: '#/definitions/PollTemplate'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/getAllPolls:
    post:
      description: Gives all the template polls
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: template poll
          in: body
          required: true
          description: poll details
          schema:
            type: object
            properties:
              last_id:
                type: string
                description: Id of the last record on the page. None if first page
              first_page:
                type: string
                description: First, previous or next
              number_of_records:
                type: number
                description: Number of records on each page
              filter_criteria:
                type: object
                properties:
                  search_value:
                    type: string
                    description: Search the template polls by title of the poll
                  category_value:
                    type: string
                    description: Search the template polls by category
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                properties:
                  polls:
                    type: array
                    items:
                      $ref: '#/definitions/PollTemplate'
                  count:
                    type: number
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/allSurveys:
    get:
      description: Gives all the template surveys
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                properties:
                  surveys:
                    type: array
                    items:
                      $ref: '#/definitions/SurveyTemplate'
                  surveyQuestions:
                    type: array
                    items:
                      $ref: '#/definitions/SurveyQuestion'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/getAllSurveys:
    post:
      description: Gives all the template surveys
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: survey
          in: body
          required: true
          description: survey details
          schema:
            type: object
            properties:
              last_id:
                type: string
                description: Id of the last record on the page. None if first page
              first_page:
                type: string
                description: First, previous or next
              number_of_records:
                type: number
                description: Number of records on each page
              filter_criteria:
                type: object
                properties:
                  search_value:
                    type: string
                    description: Search the template surveys by title of the survey
                  category_value:
                    type: string
                    description: Search the surveys by category
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                properties:
                  surveys:
                    type: array
                    items:
                      $ref: '#/definitions/SurveyTemplate'
                  surveyQuestions:
                    type: array
                    items:
                      $ref: '#/definitions/SurveyQuestion'
                  count:
                    type: number
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/allBroadcasts:
    get:
      description: Gives all the template broadcasts
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: array
                items:
                  $ref: '#/definitions/BroadcastTemplate'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/getAllBroadcasts:
    post:
      description: Gives all the template broadcasts
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: broadcast
          in: body
          required: true
          description: broadcast details
          schema:
            type: object
            properties:
              last_id:
                type: string
                description: Id of the last record on the page. None if first page
              first_page:
                type: string
                description: First, previous or next
              number_of_records:
                type: number
                description: Number of records on each page
              filter_criteria:
                type: object
                properties:
                  search_value:
                    type: string
                    description: Search the template broadcast by title of the broadcast
                  category_value:
                    type: string
                    description: Search the broadcasts by category
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                properties:
                  broadcasts:
                    type: array
                    items:
                      $ref: '#/definitions/BroadcastTemplate'
                  count:
                    type: number
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/allCategories:
    get:
      description: Gives all the categories
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: array
                items:
                  $ref: '#/definitions/Category'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/surveyDetails/:surveyid:
    get:
      description: Gives a particular survey detail
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                properties:
                  survey:
                    $ref: '#/definitions/SurveyTemplate'
                  questions:
                    type: array
                    items:
                      $ref: '#/definitions/SurveyQuestion'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/pollDetails/:pollid:
    get:
      description: Gives a particular poll template details
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/PollTemplate'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
  /templates/broadcastDetails/:broadcastid:
    get:
      description: Gives a particular poll template details
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/BroadcastTemplate'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
  /templates/createBroadcast:
    post:
      description: creates a broadcast
      consumes:
          - application/json
          - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: broadcast
          in: body
          description: broadcast to save
          schema:
            $ref: '#/definitions/BroadcastTemplate'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/BroadcastTemplate'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/editBroadcast:
    post:
      description: creates a broadcast
      consumes:
          - application/json
          - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: broadcast
          in: body
          description: broadcast to edit
          schema:
            $ref: '#/definitions/BroadcastTemplateEdit'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/BroadcastTemplateEdit'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /templates/createCategory:
    post:
      description: creates a category
      consumes:
          - application/json
          - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: category
          in: body
          description: category to save
          schema:
            $ref: '#/definitions/Category'
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/Category'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
  /templates/editCategory:
    post:
      description: edits a category
      consumes:
          - application/json
          - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: category
          in: body
          description: category to edit
          schema:
            $ref: '#/definitions/CategoryEdit'
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: object
                $ref: '#/definitions/CategoryEdit'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
...
  /templates/deleteCategory/:id:
    delete:
      description: Delete a category
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              status:
                type: string
                description: success
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
...
definitions:
...
  PollTemplate:
    type: object
    description: This table is used to store the template polls
    properties:
      _id:
        type: string
        description: unique id of the record
      title:
        type: string
        description: The title of the poll
      statement:
        type: string
        description: The statement of the poll
      category:
        type: array
        items:
          type: string
      datetime:
        type: object
        description: The date and time at which the poll is created
      options:
        type: array
        items:
          type: string
          description: The oprions of the poll
  SurveyTemplate:
    type: object
    description: This table is used to store the template surveys
    properties:
      _id:
        type: string
        description: unique id of the record
      title:
        type: string
        description: The title of the survey
      description:
        type: string
        description: The description of the survey
      category:
        type: array
        items:
          type: string
      datetime:
        type: object
        description: The date and time at which the survey is created
  SurveyQuestion:
    type: object
    description: This table is used to store the template surveys
    properties:
      _id:
        type: object
        properties:
          statement:
            type: string
            description: The statement of the survey question
          options:
            type: array
            items:
              type: string
              description: The options of the question
          surveyId:
            type: string
            description: The id of the survey
  BroadcastTemplate:
    type: object
    description: This table is used to store the template broadcasts
    required:
      - title
      - category
      - payload
    properties:
      _id:
        type: string
        description: unique id of the record
        readOnly: true
      title:
        type: string
        description: The title of the broadcast
      category:
        type: array
        items:
          type: string
      payload:
        type: array
        items:
          type: object
          description: the content of the broadcast
      datetime:
        type: object
        description: The date and time at which the broadcast is created
      userId:
        type: string
        readOnly: true
      createdBySuperUser:
        type: boolean
      companyId:
        type: string
        readOnly: true
  BroadcastTemplateEdit:
    type: object
    description: This table is used to store the template broadcasts
    properties:
      _id:
        type: string
        description: unique id of the record
        readOnly: true
      title:
        type: string
        description: The title of the broadcast
      category:
        type: array
        items:
          type: string
      payload:
        type: array
        items:
          type: object
          description: the content of the broadcast
      datetime:
        type: object
        description: The date and time at which the broadcast is created
      userId:
        type: string
        readOnly: true
      companyId:
        type: string
        readOnly: true
      createdBySuperUser:
        type: boolean
...
  Category:
    type: object
    description: This table is used to store categories for templates
    required:
      - name
    properties:
      _id:
        type: string
        description: unique id of the record
        readOnly: true
      name:
        type: string
        description: name of the category
      userId:
        type: string
        readOnly: true
      companyId:
        type: string
        readOnly: true
      createdBySuperUser:
        type: boolean
        readOnly: true
  CategoryEdit:
    type: object
    description: This table is used to store categories for templates
    required:
      - name
      - _id
    properties:
      _id:
        type: string
        description: unique id of the record
      name:
        type: string
        description: name of the category
      userId:
        type: string
        readOnly: true
      companyId:
        type: string
        readOnly: true
      createdBySuperUser:
        type: boolean
        readOnly: true
...
  Response500:
    type: object
    properties:
      status:
        type: string
        description: failed
      description:
        type: string
        description: Internal server error
...
  Response404:
    type: object
    properties:
      status:
        type: string
        description: failed
      description:
        type: string
        description: Record not found
...